Add files via upload
Notebook 01: Removed the import sys and displaying the directory - not necessary for showcasing. 
Notebook02:  Created this notebook with the structure and an introduction markdown. The data and libraries are imported and we have med a copy of the data. Ready to begin the data cleaning and feature engineering phase of the project.