Add files via upload
Removed some of steps in the data preprocessing file that were unnecessary, redid some logic in the cleaning and imputation part - there was a mistake where the features that were going to be engineered were dropped before the imputation resulting in NaN values in these - Fixed. Also some errors in the logic for the splitting as well as adding the target variable to the columns that were dropped. Lastly, error since some values in the data were not handled correctly - introducing a subsection in section 2 that handles anomalies, illogical data and inconsistencies as well as error-codes given in the code. This part is not done yet, some functions have been added and some logic exists but needs more work.